:recycle: Refactor: 그룹 관련 도메인 이동
<body>
- group/domain : Moim- class
- group/repository : repository
- grpup/type : enum 등의 type class

<footer>
- 관련: #145
