+++ application/external-api/src/main/java/com/namo/spring/application/external/api/user/facade/UserFacade.java
@@ -21,41 +21,41 @@ import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.namo.spring.application.external.api.user.converter.UserConverter;
-import com.namo.spring.application.external.api.user.converter.UserResponseConverter;
-import com.namo.spring.application.external.api.user.service.UserService;
-import com.namo.spring.application.external.domain.group.application.impl.MoimAndUserService;
-import com.namo.spring.application.external.domain.group.application.impl.MoimMemoLocationService;
-import com.namo.spring.application.external.domain.group.application.impl.MoimScheduleAndUserService;
-import com.namo.spring.application.external.domain.group.domain.MoimScheduleAndUser;
+import com.namo.spring.application.external.api.group.service.MoimAndUserService;
+import com.namo.spring.application.external.api.group.service.MoimMemoLocationService;
+import com.namo.spring.application.external.api.group.service.MoimScheduleAndUserService;
 import com.namo.spring.application.external.api.individual.converter.CategoryConverter;
 import com.namo.spring.application.external.api.individual.service.AlarmService;
 import com.namo.spring.application.external.api.individual.service.CategoryService;
 import com.namo.spring.application.external.api.individual.service.ImageService;
 import com.namo.spring.application.external.api.individual.service.PaletteService;
 import com.namo.spring.application.external.api.individual.service.ScheduleService;
+import com.namo.spring.application.external.api.user.converter.TermConverter;
+import com.namo.spring.application.external.api.user.converter.UserConverter;
+import com.namo.spring.application.external.api.user.converter.UserResponseConverter;
+import com.namo.spring.application.external.api.user.dto.UserRequest;
+import com.namo.spring.application.external.api.user.dto.UserResponse;
+import com.namo.spring.application.external.api.user.service.UserService;
+import com.namo.spring.application.external.global.common.constant.FilePath;
+import com.namo.spring.application.external.global.utils.FileUtils;
+import com.namo.spring.application.external.global.utils.JwtUtils;
+import com.namo.spring.application.external.global.utils.SocialUtils;
+import com.namo.spring.client.social.apple.client.AppleAuthClient;
+import com.namo.spring.client.social.apple.dto.AppleResponse;
+import com.namo.spring.client.social.apple.dto.AppleResponseConverter;
+import com.namo.spring.client.social.common.properties.AppleProperties;
+import com.namo.spring.client.social.common.utils.AppleUtils;
+import com.namo.spring.client.social.kakao.client.KakaoAuthClient;
+import com.namo.spring.client.social.naver.client.NaverAuthClient;
+import com.namo.spring.db.mysql.domains.group.domain.MoimScheduleAndUser;
 import com.namo.spring.db.mysql.domains.individual.domain.Category;
 import com.namo.spring.db.mysql.domains.individual.domain.Image;
 import com.namo.spring.db.mysql.domains.individual.domain.Schedule;
 import com.namo.spring.db.mysql.domains.individual.type.CategoryKind;
-import com.namo.spring.application.external.api.user.converter.TermConverter;
 import com.namo.spring.db.mysql.domains.user.domain.Term;
 import com.namo.spring.db.mysql.domains.user.domain.User;
 import com.namo.spring.db.mysql.domains.user.type.SocialType;
 import com.namo.spring.db.mysql.domains.user.type.UserStatus;
-import com.namo.spring.application.external.api.user.dto.UserRequest;
-import com.namo.spring.application.external.api.user.dto.UserResponse;
-import com.namo.spring.application.external.global.common.constant.FilePath;
-import com.namo.spring.client.social.apple.client.AppleAuthClient;
-import com.namo.spring.client.social.common.properties.AppleProperties;
-import com.namo.spring.client.social.apple.dto.AppleResponse;
-import com.namo.spring.client.social.apple.dto.AppleResponseConverter;
-import com.namo.spring.client.social.common.utils.AppleUtils;
-import com.namo.spring.client.social.kakao.client.KakaoAuthClient;
-import com.namo.spring.client.social.naver.client.NaverAuthClient;
-import com.namo.spring.application.external.global.utils.FileUtils;
-import com.namo.spring.application.external.global.utils.JwtUtils;
-import com.namo.spring.application.external.global.utils.SocialUtils;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
