--- application/external-api/src/main/java/com/namo/spring/application/external/domain/group/domain/constant/MoimScheduleStatus.java
@@ -1,5 +1,0 @@
-package com.namo.spring.application.external.domain.group.domain.constant;
-
-public enum MoimScheduleStatus {
-	ACTIVE, DELETED
-}
