--- application/external-api/src/main/java/com/namo/spring/application/external/domain/group/domain/constant/MoimStatus.java
@@ -1,5 +1,0 @@
-package com.namo.spring.application.external.domain.group.domain.constant;
-
-public enum MoimStatus {
-	INACTIVE, ACTIVE;
-}
