--- application/external-api/src/main/java/com/namo/spring/application/external/domain/group/repository/diary/MoimMemoRepository.java
@@ -1,32 +1,0 @@
-package com.namo.spring.application.external.domain.group.repository.diary;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
-import com.namo.spring.application.external.domain.group.domain.MoimMemo;
-import com.namo.spring.application.external.domain.group.domain.MoimSchedule;
-
-public interface MoimMemoRepository extends JpaRepository<MoimMemo, Long> {
-	@Query(value = "select distinct mm"
-		+ " from MoimMemo mm"
-		+ " join fetch mm.moimSchedule ms"
-		+ " join fetch ms.moimScheduleAndUsers msu" + " join fetch msu.user" + " where ms = :moimSchedule")
-	MoimMemo findMoimMemoAndUsersByMoimSchedule(@Param("moimSchedule") MoimSchedule moimSchedule);
-
-	Optional<MoimMemo> findMoimMemoByMoimSchedule(MoimSchedule moimSchedule);
-
-	@Query("select mm"
-		+ " from MoimMemo mm"
-		+ " left join fetch mm.moimMemoLocations"
-		+ " where mm.moimSchedule =:moimSchedule")
-	MoimMemo findMoimMemoAndLocationsByMoimSchedule(MoimSchedule moimSchedule);
-
-	@Query("select mm"
-		+ " from MoimMemo mm"
-		+ " left join fetch mm.moimMemoLocations"
-		+ " where mm.moimSchedule.id =:moimScheduleId")
-	Optional<MoimMemo> findMoimMemoAndLocationsByMoimSchedule(Long moimScheduleId);
-}
