+++ storage/db-mysql/src/main/java/com/namo/spring/db/mysql/domains/group/repository/group/MoimRepositoryCustom.java
@@ -1,0 +1,9 @@
+package com.namo.spring.db.mysql.domains.group.repository.group;
+
+import com.namo.spring.db.mysql.domains.group.domain.Moim;
+
+public interface MoimRepositoryCustom {
+	Moim findMoimHavingLockWithMoimAndUsersByCode(String code);
+
+	Moim findHavingLockById(Long moimId);
+}
