+++ storage/db-mysql/src/main/java/com/namo/spring/db/mysql/domains/group/type/MoimStatus.java
@@ -1,0 +1,5 @@
+package com.namo.spring.db.mysql.domains.group.type;
+
+public enum MoimStatus {
+	INACTIVE, ACTIVE;
+}
